chore(web-ui): rename assets
