+++ kagami/webui/src/page/Ad/Announcement/styles.module.scss
@@ -35,7 +35,7 @@
               // 背景颜色
     box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1); // 阴影效果
     position:relative;
-    background-image: url('../../../picture/kagami.jpg'); /* 背景图片 */
+    background-image: url('../../../assets/kagami.png'); /* 背景图片 */
     background-size: cover; /* 背景图片覆盖整个容器 */
     background-position: center; /* 背景图片居中 */
     background-repeat: no-repeat; /* 背景图片不重复 */
